xmpp.go: Be more permissive when doing the initial SRV request.
Some jabber servers use subdomains like `jabber.nullcat.de`.  I think it's valid
when only the top-level domain (nullcat.de) would have an SRV record, but not
the full one. This leads to a failure in xmpp.NewClient().

This patch goes over the domain and shortens it on each retry.